requirements.txt
numpy==1.23.5
pandas==1.5.2
scikit-learn==1.3.0
scipy==1.10.0
matplotlib==3.6.2
tqdm
ujson >= 4.0.2
wandb
pyarrow==2.0.0
pytorch-lightning==1.1.4
transformers==4.2.1
timm
kaggle